fixed bad version error where there were multiple files
